arbol

+++ app/src/main/java/com/example/garya/lab2android/Caracter.java
@@ -15,7 +15,7 @@ public class Caracter {
         this.caracter = caracter;
     }
 
-    public double getProbabilidad() {
+    public float getProbabilidad() {
         return probabilidad;
     }
